Add language from database to est and ms text endpoint responses
If the language of an established text or a manuscript is defined in the database, include the language value in the endpoint response. This is needed for the frontend to correctly add lang-attributes to the HTML in case the texts are in a different language than the user interface.

This commit also removes
        if language is not None:
            data["language"] = language
from the est endpoint because there is no need to include the request language in the response.