fix: set CORS header for `get_pdf_file` endpoint
Dynamically set the Access-Control-Allow-Origin header to the request origin if the origin is a *.sls.fi subdomain. This makes it possible to embed PDFs served by the API on these sites.

Refactor if-elif code block setting `file_path` so there is only one branch per file type, and add mimetype to `send_file`.

Place the `connection.close()` call within a `finally` block to ensure connection closure on early returns.