fix: set CORS header for `get_pdf_file` endpoint (#39)
Dynamically set the Access-Control-Allow-Origin header to the request origin, if the origin is allowed. Allowed origins are defined in a list with the key `allowed_cors_origins` in the project sections of the config YAML file. This makes it possible to embed PDFs served by the API on these sites.

The code is backwards compatible with config files where the `allowed_cors_origins` key is missing in project sections.

Additional changes in the `get_pdf_file` endpoint:
 - Refactor if-elif code block setting `file_path` so there is only one branch per file type, and add mimetype to `send_file`.
 - Place the `connection.close()` call within a `finally` block to ensure connection closure on early returns.