fix: set CSP frame-ancestors header to allow embedding PDFs from specified sources
This fixes an issue where PDFs served by the API can’t be embedded in webpages. Previously this was erraneously attempted to be fixed by setting CORS headers on the response.

This approach sets the Content-Security-Policy header with the frame-ancestors directive to host sources specified in the project config, to allow embedding PDFs served by the API on these specific hosts.

TODO: ensure that the older X-Frame-Options header is not set by nginx or a reverse proxy on the response, if the response contains a Content-Security-Policy: frame-ancestors header.

The X-Frame-Options and CSP frame-ancestors headers are intended to be used independently and both control framing behaviour.

MDN docs on the headers:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/frame-ancestors
https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Frame-Options